+++ event/queue/errors.go
@@ -1,4 +1,4 @@
-package eventqueue
+package queue
 
 import (
 	"net/http"
